requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.2
anyio==4.12.1
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
astunparse==1.6.3
async-lru==2.2.0
async-timeout==5.0.1
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
dask==2026.1.2
debugpy==1.8.20
decorator==5.2.1
DeepXDE==1.15.0
defusedxml==0.7.1
distlib==0.4.0
dm-tree==0.1.9
exceptiongroup==1.3.1
executing==2.2.1
Farama-Notifications==0.0.4
fastjsonschema==2.21.2
filelock==3.24.3
fire==0.7.1
flatbuffers==25.12.19
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.12.0
future==1.0.0
gast==0.7.0
gitapi==1.0.1
google-pasta==0.2.0
greenlet==3.3.2
grpcio==1.78.0
gymnasium==1.2.3
h11==0.16.0
h5py==3.15.1
hgapi==1.7.4
httpcore==1.0.9
httpx==0.28.1
hyperopt==0.2.7
identify==2.6.16
idna==3.11
importlib_metadata==8.7.1
ipykernel==7.2.0
ipython==8.38.0
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
keras==3.12.1
kiwisolver==1.4.9
lark==1.3.1
lark-parser==0.12.0
libclang==18.1.1
lightning-utilities==0.15.3
llvmlite==0.46.0
locket==1.0.0
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
meep==1.0.6
mistune==3.2.0
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.1
namex==0.1.0
narwhals==2.17.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.10.0
notebook==7.5.4
notebook_shim==0.2.4
numba==0.64.0
numpy==2.2.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opt_einsum==3.4.0
optree==0.19.0
optuna==4.7.0
overrides==7.7.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.6
partd==1.4.2
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.9.2
plotly==6.5.2
pre_commit==4.5.1
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==7.34.0
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
pyaml==26.2.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyepr==1.3.0
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-discovery==1.1.0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytorch-lightning==2.6.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
quantum-blackbird==0.5.0
quantum-xir==0.2.2
qutip==5.2.3
qutip-qip==0.4.1
qutip-qtrl==0.1.5
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rpds-py==0.30.0
sb3_contrib==2.7.1
scikit-learn==1.7.2
scikit-optimize==0.10.2
scikit-rf==1.10.0
scipy==1.15.3
seaborn==0.13.2
Send2Trash==2.1.0
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.47
stable_baselines3==2.7.1
stack-data==0.6.3
StrawberryFields==0.23.0
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-probability==0.25.0
termcolor==3.3.0
terminado==0.18.1
thewalrus==0.22.0
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.4.0
toolz==1.1.0
torch==2.5.1+cu121
torch-geometric==2.7.0
torch_cluster==1.6.3+pt25cu121
torch_scatter==2.1.2+pt25cu121
torch_sparse==0.6.18+pt25cu121
torch_spline_conv==1.2.2+pt25cu121
torchaudio==2.5.1+cu121
torchmetrics==1.8.2
torchvision==0.20.1+cu121
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
triton==3.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
virtualenv==21.1.0
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.6
widgetsnbextension==4.0.15
wrapt==2.1.1
xanadu-cloud-client==0.3.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
